nftables: implement non-loose reverse path filtering
IPv6_rpfilter is documented to accept packets "If a reply to the packet
would be sent via the same interface that the packet arrived on". The
ip6tables backend implements this behaviour as it does not use the --loose
option with the rpfilter match.

The nftables backend, however uses a fib expression that will match a
packet with a source address that has a valid route, regardless of the
interface the route would use.

The fib expression adopted by this change is taken from the nftables man
page and is documented to "drop packets to address not configurated on
ininterface".